requirements.txt
pandas==2.2.3
PyPDF2==3.0.1
boto3==1.28.75
azure-storage-blob==12.17.0